Add preliminary work to support Babbler-based apps
Amazon, for examples, still refuses to add chromecast support, but there
are a number of people who have documented how to work with their DRM
stuff. Babbler is a custom Chromecast receiver that people can set up
that enables you to play DASH media using a side-channel for the license
requests. Taking advantage of both of those, we can create Chromecast
support for Amazon Prime videos where there was none before!

To be very clear, none of this is an attempt to subvert Amazon's DRM,
and indeed, like all the other Apps on here, you need a valid account
with valid cookies to use this. All we're doing is enabling convenient,
programmatic playback of content that you already have access to.