Begin work to support auto-updating credentials
Refs #8

Includes support for HboGo so far. Hulu has never gotten logged out, but
it may be worthwhile to add support to Youtube; we just need to respect
the Set-Cookie headers, presumably.

Using a wrapper type for Tokens we can update tokens as we go if we're
provided an `IWritableToken` instance instead of just a `string`, which
the `Player` interface will always do for us.

If you're manually opening an app without `Player` then you're
on your own for providing an `IWritableToken` implementation....