Add support for "real" Prime video casting (#10)
* Rename old Babbler-based Prime video app

* WIP experiments with using Amazon's new chromecast support

Not quite sure how to compute the preAuthorizedLinkCode to authorize
the session, and definitely not sure how to get the title IDs. Maybe
if they add support to the website we can find out something more
easily...

* Minor refactoring

* Support generating a link code for Prime video

Still missing a step somewhere, because Register now throws an empty
object (`{}`) error. Probably something to do with the weird
AmIRegistered handshake?

* Experiment with generating FRC cookies; add cli to get token

Still getting an empty error map from the chromecast :\  No idea
what step we're missing here....

* Support providing deviceType to for iterative testing

* Adjust device_model + /code/create calls based on experimental data

Now, with a "proper" refresh token, we can successfully register
and cast videos to the chromecast! Woohoo!!!

Now the tricky part will probably be *acquiring* that refresh token....

* Separate out Prime API + add manual Prime auth!

Woohoo! We now have a complete loop!

* Store web cookies and use with Chakram for various Player APIs

It'd be nice to use the real APIs instead of web scraping, but
I haven't found any that directly do what we want, so this is
fine for now....

* Drop unused oauth-based Prime auth

* Refactor FRC cookie generation; strip references to the Java code