Add a convoluted mechanism for resuming prime video series
See #13

This is probably not a 100% solution, and is incompatible with
playing URLs (grumble grumble) but it *should* work fairly well
for resuming search results. Will also probably need to add extra
handling for movies.

The `getTitleInfo` endpoint does have an "aliases" field that we
may be able to use to match up the ASINs from the URLs...