Add ChromecastDevice.detect() method
v1.19.0

This will return the device's model, among other things, which can be
useful if you're preparing video to play on the device and need to check
its capabilities, first. The model can be cross-referenced with the
documentation here: https://developers.google.com/cast/docs/media

This commit includes a refactor to import device detection into this
library; it is almost completely based on the original in nodecastor
so there should be no change in behavior.