Revert: skipIntro PlayableOptions
The behavior of these would be pretty weird: unlike Netflix, where
it will never auto-skip intros on initial play but *might* auto-skip
when continuing to "next," this *might* auto-skip on intiail play, but
could *never* when continuing to "next."

Sadly, we must drop it.

It may be possible to implement an explicit command to perform a skip
based on the current playback timestamp, but that would be unrelated to
these options.