prime: Improve reliability of watchNext and queue resolution
The "Watch Next" carousel is unfortunately not always on the first page
of the home collections resources, so we have to paginate that to ensure
we can get "Watch Next." I probably mentioned this in a previous commit,
but we can't actually just rely on the `selectedEpisode`, because it
is specific to the *season*, and does not work across seasons, so it's
effectively useless in a multi-season series.

Next, we had to ensure that the ITitleInfo we pass to the resolveQueue
method is actually for the season that the upNext episode belongs to, in
order to correctly fetch queue data.