Update HBO to support latest protocol changes (#42)
* hbo: Improve load error message

* Attempt to update HBO protocol to match recent changes

Even if I send exactly the same thing that the web app appears to be
sending, HBO claims that the content isn't included in the subscription.
Very strange and frustrating....

* Update HBO protocol to use the custom LOAD_MEDIA message

Who knows why HBO did this but it's now much more annoying to detect
that the cast was accepted (and it's also... well, non-standard). But
they did, and this seems to work.