Export the Player type; add buildQueryOnly + buildUpon
`buildQueryOnly()` lets us construct Player instances that we intend to
use only for query purposes without having to provide a bogus
ChromecastDevice just to satisfy the requirements to `build()`. The
return type of this new method reflects that it does not have the
ability to `play()` anything.

In addition, we now provide `Player.buildUpon` to construct a new
`PlayerBuilder` with the same starting params as the Player,
facilitating a factory-like approach where you eg: reuse configured apps
then `buildUpon().addDevice(...).build()` to play on a specified device
per-request.