Add new queryRecommendations interface; deprecate queryRecommended
All queryRecommended methods should now return what they returned before
this branch; we will not change its behavior since that would be
unnecessarily breaking. Instead, we introduce the new
`queryRecommendations` method which is explicitly labeled as unstable
for now.

Most channels still return "recents" for this method, which is fine.